Update project cracker runtime detection method
The downloadDependencies method doesn't seem to work on .NET Core 3.1,
which doesn't include that information for the basic framework. That
means that System.Core.dll etc. can still go unreferenced.

Instead, we can check the frameworkReferences to get a list of frameworks
like ["Microsoft.NETCore.App", "Microsoft.AspNetCore.App"] and resolve
these to full paths using dotnet --info. That gives the same set of
assemblies as the downloadDependencies method, but with a few advantages:

- frameworkReferences is common across both different dotnet versions
  and on different platforms

- We can support ASP.NET Core projects and properly reference their
  assemblies in addition to the ones that come with the base framework