Update framework detection to account for backwards compatibility
.NET Core lets you run netcoreapp3.x as long as you have some .NET Core 3.y
installed, as long as y>=x. That means that we should only reject frameworks if
that either differ by major version or are not of a high enough minor version.

To avoid conflicts, we try to resolve frameworks so that the minor version is as
close as possible to the target framework. This reduces the chance of incorrect
suggestions popping up due to API additions (classes/members which exist in .NET
Core 3.y that aren't there in .NET Core 3.x).