Support .NET Core 3 projects (#70)
* Migrate to netcoreapp3.0 and support its projects

The main changes required for netcoreapp3.0 are due to
https://github.com/NuGet/Home/wiki/%5BSpec%5D-PackageDownload-support.
This affects us because it means that the framework dependencies are no
longer included directly in projects.assets.json. Instead we have to
look for them within the dotnet install directory.

The basic process is:

- Figure out what framework applies by looking at downloadDependencies
  within the project.assets.json

- Invoke dotnet --info to find the base path of the corresponding
  framework

- Reference all the DLLs in that directory

* Add test for netcoreapp3 implied dependencies

* Don't crash if we can't find a runtime directory

* Update VS Code extension files to use netcoreapp3.0

* Update project cracker runtime detection method

The downloadDependencies method doesn't seem to work on .NET Core 3.1,
which doesn't include that information for the basic framework. That
means that System.Core.dll etc. can still go unreferenced.

Instead, we can check the frameworkReferences to get a list of frameworks
like ["Microsoft.NETCore.App", "Microsoft.AspNetCore.App"] and resolve
these to full paths using dotnet --info. That gives the same set of
assemblies as the downloadDependencies method, but with a few advantages:

- frameworkReferences is common across both different dotnet versions
  and on different platforms

- We can support ASP.NET Core projects and properly reference their
  assemblies in addition to the ones that come with the base framework

* Fix runtime detection to handle paths with spaces

* Bad syntax

* Add fallback for project files which lack a frameworkReferences

* Code cleanup before PR

* Update Appveyor config to use Windows build with .NET Core 3

* Add netcoreapp3.1 as a supported framework

* Update framework detection to account for backwards compatibility

.NET Core lets you run netcoreapp3.x as long as you have some .NET Core 3.y
installed, as long as y>=x. That means that we should only reject frameworks if
that either differ by major version or are not of a high enough minor version.

To avoid conflicts, we try to resolve frameworks so that the minor version is as
close as possible to the target framework. This reduces the chance of incorrect
suggestions popping up due to API additions (classes/members which exist in .NET
Core 3.y that aren't there in .NET Core 3.x).

* Allow framework discovery to consider both packs and runtimes

Packs are better tailored to a specific platform, and don't include
assemblies which can't be loaded on it. Runtimes can include unloadable
assemblies that cause completion failures when a member refers to some
assembly that can't be loaded.

* Add support for loading framework packs downloaded via NuGet

dotnet restore will install these packages if the runtime framework is newer
than the one referenced by the project. We should consider them if we can find
them to get a version that as close to possible as the target.

* Fix merge conflict

* Trigger rebuild

* Switch c# dep

* Update instructions

* Ignore other versions

* Update other paths

Co-authored-by: George Fraser <george@fivetran.com>